Changed ClipImage Component
Clip image now uses a <div> element to display the image instead of an <img> element. This was done to be able to have the image stretch over the available screen space on the home page.

Also made the className available as a property.

Added a few test element in the readme.md